Improving the performance of the code. Now the data buffer is created only one time and it is reused, avoiding unnecessary malloc overhead.
Bug fixed when reading data from a digital channel, now is considering that the position of the bit line is not located in the bit 0 of the UINT32.

Bug fixed in the plot buffer, now is using the np.float64 data type instead of float.

Introducing a debug mode to avoid the exp_dir, tribu_id and rload_id each time that we are running the code.